Create Readme.md
DESCRIPTION

Tic Tac Toe is a game where two players on the other hand put Xs and Os in compartments of a figure framed by two vertical lines crossing two flat lines and each attempts to get a line of three Xs or three Os before the rival does. Spasm tac-toe, noughts and crosses, or Xs and Os is a paper-and-pencil game for two players who alternate denoting the spaces in a three-by-three matrix with X or O. The player who prevails with regards to putting three of their imprints in a flat, vertical, or slanting column is the victor. It is an addressed game, with a constrained draw expecting best play from the two players.

-> Prerequisites

Undeniable LEVEL REQUIREMENTS
On the off chance that both the players play ideally, it is ordained that you won't ever lose ("albeit the match can in any case be drawn"). It doesn't make any difference whether you play first or second.In another ways - " Two master players will continuously draw ". As said above, on the off chance that two experienced players are playing the Tic-Tac-Toe, the game will continuously draw. There is one more popular variation of this game-Ultimate Tic-Tac-Toe, which means to make the ordinary Tic-Tac-Toe really intriguing and less unsurprising.

LOW LEVEL REQUIREMENTS :

-> SWOT ANALYSIS :

Qualities : Below are the Strengths in the SWOT Analysis of Tic Tac:
Shortcoming :
Amazing open doors :
Danger :
-> 4W and 1H

Who : Any Person ready to play aganist PC

What : Play Tic Tac Toe Game utilizing C programming in Visual studio Code

Where : Online or on Visual studio code or on any C compiler which can run a multifile program without any problem

When : Anytime

How : Just Open the github repository in Visual Studio code or clone the repository and